* generic/tkFileFilter.c (AddClause): Cast to match function prototype. 	* win/stubs.c (_XInitImageFuncPtrs): Add return value for function. 	* win/tkWinButton.c (buttonStyles, ButtonBindProc, ComputeStyle): 	Remove unused declarations. 	* win/tkWinColor.c (GetColorByName, GetColorByValue): Remove unused 	function declarations. 	* win/tkWinDialog.c (TrySetDirectory): Remove unused function declaration. 	* win/tkWinEmbed.c (TkWinEmbeddedEventProc): Cast to match function prototype. 	* win/tkWinMenu.c (winMenuMutex, MenuExitProc): Remove unused declaration. 	* win/tkWinWindow.c (StackWindow): Remove unused declaration. 	* win/tkWinWm.c (ConfigureEvent): Remove unused declaration. 	* win/tkWinX.c (winXMutex): Remove unused declaration. 	* xlib/ximage.c (XCreateBitmapFromData): Cast to match function prototype.

+++ xlib/ximage.c
@@ -8,7 +8,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: ximage.c,v 1.3 1998/09/14 18:24:03 stanton Exp $
+ * RCS: @(#) $Id: ximage.c,v 1.4 2000/07/06 03:17:45 mo Exp $
  */
 
 #include "tkInt.h"
@@ -47,7 +47,7 @@ XCreateBitmapFromData(display, d, data, width, height)
     GC gc;
     Pixmap pix;
 
-    pix = Tk_GetPixmap(display, d, width, height, 1);
+    pix = Tk_GetPixmap(display, d, (int) width, (int) height, 1);
     gc = XCreateGC(display, pix, 0, NULL);
     if (gc == NULL) {
 	return None;
